requirements.txt
asgiref==3.9.1
Django==5.2.6
django-nested-admin==4.1.4
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
openpyxl==3.1.5
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-monkey-business==1.1.0
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.10.0
